+++ mod/booking/classes/option/fields/addtogroup.php
@@ -83,7 +83,7 @@ class addtogroup extends field_base {
      * @param stdClass $formdata
      * @param stdClass $newoption
      * @param int $updateparam
-     * @param mixed $returnvalue
+     * @param ?mixed $returnvalue
      * @return string // If no warning, empty string.
      */
     public static function prepare_save_field(
@@ -122,7 +122,7 @@ class addtogroup extends field_base {
 
             $bookingsettings = singleton_service::get_instance_of_booking_settings_by_cmid($cmid);
 
-            if (!empty($bookingsettings->addtogroup) && $option->courseid > 0) {
+            if (!empty($bookingsettings->addtogroup) && !empty($option->courseid)) {
                 $bo = singleton_service::get_instance_of_booking_option($cmid, $optionid);
                 // TODO: This looks kind of strange. Was this copied from legacy code? Does it still work?
                 // phpcs:ignore Squiz.PHP.CommentedOutCode.Found