+++ mod/booking/classes/signinsheet/signinsheet_generator.php
@@ -221,10 +221,11 @@ class signinsheet_generator {
     /**
      * Define basic variable values for signinsheet pdf
      *
-     * @param \mod_booking\booking_option $bookingoption
      * @param \stdClass $pdfoptions
+     * @param ?\mod_booking\booking_option $bookingoption
+     *
      */
-    public function __construct(\mod_booking\booking_option $bookingoption = null, \stdClass $pdfoptions) {
+    public function __construct(\stdClass $pdfoptions, ?\mod_booking\booking_option $bookingoption = null) {
 
         $this->optionid = $bookingoption->optionid;
         $this->bookingoption = $bookingoption;