+++ mod/booking/classes/task/send_mail_by_rule_adhoc.php
@@ -105,11 +105,12 @@ class send_mail_by_rule_adhoc extends \core\task\adhoc_task {
             try {
                 // Use message controller to send the message.
                 $messagecontroller = new message_controller(
-                    MOD_BOOKING_MSGCONTRPARAM_SEND_NOW, MOD_BOOKING_MSGPARAM_CUSTOM_MESSAGE,
+                    MOD_BOOKING_MSGCONTRPARAM_SEND_NOW,
+                    MOD_BOOKING_MSGPARAM_CUSTOM_MESSAGE,
                     $taskdata->cmid,
-                    null,
                     $taskdata->optionid,
                     $taskdata->userid,
+                    null,
                     null,
                     null,
                     $taskdata->customsubject,
@@ -117,6 +118,7 @@ class send_mail_by_rule_adhoc extends \core\task\adhoc_task {
                     $taskdata->installmentnr ?? 0,
                     $taskdata->duedate ?? 0,
                     $taskdata->price ?? 0,
+                    $taskdata->rulejson ?? 0,
                 );
             } catch (Exception $e) {
                 if (get_config('booking', 'bookingdebugmode')) {