+++ mod/booking/settings.php
@@ -346,6 +346,7 @@ if ($ADMIN->fulltree) {
             [],
             $options));
 
+        // phpcs:disable
         // $settings->add(
         //     new admin_setting_configcheckbox('booking/usecoursecategorytemplates',
         //             get_string('usecoursecategorytemplates', 'mod_booking'),
@@ -355,6 +356,7 @@ if ($ADMIN->fulltree) {
         //     new admin_setting_configtext('booking/templatecategoryname',
         //         get_string('templatecategoryname', 'mod_booking'),
         //         get_string('templatecategoryname_desc', 'mod_booking'), '', PARAM_TEXT));
+        // phpcs:enable
 
     } else {
         $settings->add(
@@ -642,7 +644,7 @@ if ($ADMIN->fulltree) {
     // PRO feature: Booking actions.
     // Booking actions are not yet finished, so we do not show them yet.
     // phpcs:ignore Squiz.PHP.CommentedOutCode.Found
-    /* if ($proversion) {
+    if ($proversion) {
         $settings->add(
             new admin_setting_heading('boactions',
                 get_string('boactions', 'mod_booking'),
@@ -656,7 +658,7 @@ if ($ADMIN->fulltree) {
             new admin_setting_heading('boactions',
                 get_string('boactions', 'mod_booking'),
                 get_string('infotext:prolicensenecessary', 'mod_booking')));
-    } */
+    }
 
     // PRO feature: Progress bars.
     if ($proversion) {